Error on errors
An appropriate error code is used for exiting during most situations, but not the final catch of the blocks above. In order to use this library in our build process, we need it to consistently error if something goes wrong. This PR is intended to make the erroring more consistent.

Alternatively, we would need the internal functions to be exposed, so we could create our own runner for labelling and generating icons. Please advise if the latter is desired, and I shall try and find time to create that pull request instead.